CSS, Favicon, Animation and Comments
All these changes are towards the index.html file.

CSS animation added to the pictures of the predator and human aswell as the logo/name.

I have added 2 favicons to the rootfolder from the logo that was made. 16x16 and 32x32.

Deleted unnessarry CSS from the CSS Grid.

Made better overview and structure with comments in the CSS.