Pictures, HTML and CSS, Animation, html to PHP
Added new pictures and deleted  old ones.

Deleted some HTML so that the new animations could work properly.

Added, deleted and structured CSS. Added new animations to the predator, the human and the logo.  Added in some of Linas HTML and CSS to the search bar, but not all yet.

Edited my index.html to index.php so that i can work with the other pages.